Added a few more things to the autoconf/automake scripts
configure.in now exits if required libraries are not found.
I added the headers to the makefile to be installed, but they're not
really required at the moment. I might add this properly later, so
for now I'll just leave them commented out.

I also added the wisdom and clig lines to the makefile